stock_memo/requirements.txt
anthropic>=0.50.0
tweepy>=4.14.0
python-dotenv>=1.0.0
yfinance>=0.2.36
requests>=2.31.0
